% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desvio_padrao.R
\name{desvio_pad}
\alias{desvio_pad}
\title{Calcula o desvio padrao para um vetor numerico}
\usage{
desvio_pad(x)
}
\arguments{
\item{x}{o objeto ou vetor numerico}
}
\value{
o desvio padrao do vetor numerico
}
\description{
Retorna a raiz quadrada da variancia
}
