% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erro_padrao.R
\name{erro_pad}
\alias{erro_pad}
\title{Calcula o erro padrao para um vetor numerico}
\usage{
erro_pad(x)
}
\arguments{
\item{x}{o objeto ou vetor numerico}
}
\value{
o erro padrao
}
\description{
Divide a variancia do vetor de entrada pelo numero de elementos desse vetor e calcula a raiz quadrada
}
