% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{media}
\alias{media}
\title{Calcula a media para um vetor numerico}
\usage{
media(x)
}
\arguments{
\item{x}{objeto ou vetor numerico}
}
\value{
valor medio
}
\description{
Soma todos os valores do vetor de entrada e divide pelo numero de elementos desse vetor
}
