% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediana.R
\name{mediana}
\alias{mediana}
\title{Calcula a mediana para um vetor numerico}
\usage{
mediana(x)
}
\arguments{
\item{x}{o objeto ou vetor numerico}
}
\value{
o valor central do vetor
}
\description{
Ordena os valores. Se o número de elementos eh par, soma os dois valores centrais e divide por 2. Se impar, a mediana eh o valor central.
}
