% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{moda}
\alias{moda}
\title{Calcula moda para um vetor numerico}
\usage{
moda(x)
}
\arguments{
\item{x}{o objeto ou vetor numerico}
}
\value{
o valor mais frequente e o numero de vezes que ocorre
}
\description{
Retorna o valor que aparece com maior frequencia em um vetor numerico
}
