% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumo.R
\name{resumo}
\alias{resumo}
\title{Sumariza diferente metricas para um vetor numerico}
\usage{
resumo(x, na.rm = TRUE)
}
\arguments{
\item{x}{o objeto ou vetor numerico}

\item{na.rm}{remover NAs; padrão TRUE}
}
\value{
o resumo estatistico
}
\description{
Calcula para um vetor numerico o numero de elementos, a media, o erro padrao e o desvio padrao
}
